pg_trigger/requirements.txt
async-timeout==4.0.2
attrs==21.4.0
black==22.1.0
branca==0.4.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
Deprecated==1.2.13
Fiona==1.8.21
flake8==4.0.1
folium==0.12.1.post1
fonttools==4.31.2
geographiclib==1.52
geojson==2.5.0
geopandas==0.9.0
geopy==2.2.0
greenlet==1.1.2
h3==3.7.3
idna==3.3
igraph==0.9.9
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
mccabe==0.6.1
mpmath==1.2.1
munch==2.5.0
mypy-extensions==0.4.3
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pathspec==0.9.0
patsy==0.5.2
Pillow==9.0.1
platformdirs==2.5.1
powerlaw==1.5
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pymongo==4.0.2
pyparsing==3.0.7
pyproj==3.3.0
python-dateutil==2.8.2
python-igraph==0.9.9
pytz==2022.1
PyYAML==6.0
redis==4.2.0
redis2==2.10.6.3
requests==2.27.1
scikit-learn==1.0.2
scikit-mobility==1.2.3
scipy==1.8.0
Shapely==1.8.1.post1
six==1.16.0
SQLAlchemy==1.4.32
statsmodels==0.13.2
tabulate==0.8.9
texttable==1.6.4
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.63.1
typing_extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.0
